backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mysqlclient
numpy
pandas
prophet
neuralprophet
sqlalchemy
scikit-learn
statsmodels
gunicorn
django-environ
autogluon